refactor : moved 'imageHandler' into utils file - reason : 'imageHandler' is method, not UI.
